Last refactoring for the header component.
